requirements.txt
boto3==1.34.34
botocore==1.34.34
certifi==2024.2.2
charset-normalizer==3.3.2
Cython==3.0.8
docopt==0.6.2
greenlet==3.0.3
idna==3.6
jmespath==1.0.1
numpy==1.24.4
pandas==2.0.3
pipreqs==0.4.13
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
yarg==0.1.9
